+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+- PDBeMolstarPlugin.extensions.foldseek
+
 ## [v3.2.0] - 2024-04-24
 - Mol* core dependency updated to 3.45.0
 - Removed Assembly Symmetry hack (now will hide assembly symmetry section for non-biological assemblies)